+++ src/robin_hood.h
@@ -481,7 +481,6 @@ private:
         size_t const numElementsToAlloc = calcNumElementsToAlloc();
 
         // alloc new memory: [prev |T, T, ... T]
-        // std::cout << (sizeof(T*) + ALIGNED_SIZE * numElementsToAlloc) << " bytes" << std::endl;
         size_t const bytes = ALIGNMENT + ALIGNED_SIZE * numElementsToAlloc;
         add(assertNotNull<std::bad_alloc>(malloc(bytes)), bytes);
         return mHead;